backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
supabase==2.28.2
python-dotenv==1.2.2
pydantic==2.12.5
httpx==0.28.1
